Moved the JavaScript files to their corresponding folders as part of a tidy-up. The HTML files have not yet been wired to their corresponding js files within the folders.
