+++ index.html
@@ -2878,7 +2878,7 @@
         </div>
     </footer>
 
-    <script src="static/js/main.js" type="module"></script>
+    <script src="static/js/pages/main.js" type="module"></script>
     <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11.12.2/dist/sweetalert2.all.min.js"></script>
