+++ static/js/handlers/itemCardInteractions.js
@@ -1,26 +1,20 @@
 import { buildQuickView, 
-         centerSubscribeText, 
          closeItemQuickView, 
-         displaySubscribedMessage, 
-         removeSubscriptionForm }  from "../builder.js";
+        }  from "../builders/builder.js";
 
 import { displayWishListMessage as handleWishListOpenMsg, 
          displayAddToCartMessage 
-        }  from "../messages.js";
+        }  from "../messages/messages.js";
 
-import { getItemFromLocalStorage as getJWtToken,
-         saveToLocalStorage      as setJWtToken
-        } from "./utils.js";
 
-import { JWT }                from "./jwtToken.js";
-import getItemByID            from "./itemUtils.js";
+
+import getItemByID            from "../utils/itemUtils.js";
 import handleImagesRotation   from './imageRotationHandler.js';
-import ItemCart               from "../cart.js";
+import ItemCart               from "../modules/cart.js";
 
 
 const cart      = new ItemCart();
 const quickView = document.getElementById("quick-view");
-
 
 
 /**
@@ -203,75 +197,17 @@ function handleCardMenuDisplayRemoval(cardMenu) {
 }
 
 
-async function handleSubscribeForm(e, subscribeForm, secret_key) {
-    e.preventDefault();
-    
-    const form         = new FormData(subscribeForm);
-    const emailAddress = form.get("email");
 
-    // Not built yet but a fetch method will be here that will subscribe to the backend
-    // by passing in the email address
-    
-    // Example fetch call (commented out until implemented):
-    // const response = await fetch('backend-endpoint', {
-    //     method: 'POST',
-    //     headers: {
-    //         'Content-Type': 'application/json',
-    //         'Authorization': `Bearer ${secret_key}`
-    //     },
-    //     body: JSON.stringify({ email: emailAddress })
-    // });
-
-    Swal.fire({
-        title: "Subscription Successful!",
-        text: "Thank you for subscribing to our site. Stay tuned for updates!",
-        icon: "success"
-    });
-
-    const jwt       = new JWT();
-    const milliseconds = 1000;
-    const payload   = {
-        email: emailAddress,
-        role: "subcribed",
-        lat:  Math.floor(Date.now() / milliseconds),  // Current UNIX timestamp
-        exp: 2145916800                               // Expires on January 1, 2038 (UNIX timestamp)
-
-    }
-    const JWT_TOKEN = await jwt.createJWTToken(payload, secret_key);
-    
-    if (JWT_TOKEN) {
-        setJWtToken("subscribed", JWT_TOKEN);
-        centerSubscribeText();
-        removeSubscriptionForm();
-        displaySubscribedMessage();
-    }
-
-
-}
-
-
-function setupEventListeners() {
-    document.addEventListener("DOMContentLoaded", () => {
-        const token = getJWtToken("subscribed");
-        if (token) {
-            removeSubscriptionForm();
-            displaySubscribedMessage();
-            centerSubscribeText();
-        }
-    });
-
-}
 
 
 
 export {
     addCardEventListeners,
-    handleSubscribeForm,
     handleQuickView,
     handleAddItemToCart,
     handleCartItemQuantityChange,
     handleClearCart,
     addImageRotationListeners,
-    setupEventListeners,
+ 
  
 }
