+++ templates/account/orders/invoice.html
@@ -141,7 +141,7 @@
 
     </section>
 
-    <script src="../../../static/js/invoice.js" type="module"></script>
+    <script src="../../../static/js/pages/invoice.js" type="module"></script>
 </body>
 
 </html>
