+++ templates/account/orders/orders.html
@@ -727,7 +727,7 @@
         </div>
     </footer>
 
-    <script src="../../../static/js/orders.js" type="module"></script>
+    <script src="../../../static/js/pages/orders.js" type="module"></script>
     <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11.12.2/dist/sweetalert2.all.min.js"></script>
