+++ templates/account/orders/view-item.html
@@ -426,7 +426,7 @@
         </div>
     </footer>
 
-    <script src="../../../static/js/view-item.js" type="module"></script>
+    <script src="../../../static/js/pages/view-item.js" type="module"></script>
     <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11.12.2/dist/sweetalert2.all.min.js"></script>
