+++ templates/account/reviews/add-review.html
@@ -370,8 +370,8 @@
         </div>
     </footer>
 
-    <script src="../../../static/js/add-review.js" type="module"></script>
-    <script src="../../../static/js/reviews.js" type="module"></script>
+    <script src="../../../static/js/modules/add-review.js" type="module"></script>
+    <script src="../../../static/js/modules/reviews.js" type="module"></script>
     <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11.12.2/dist/sweetalert2.all.min.js"></script>
